requirements.txt
flask==1.1.1
connexion==2.4.0
werkzeug==0.16.1
swagger-ui-bundle
prance==0.20.2
waitress==1.3.1
audit-middleware==0.2.7
auth-middleware==4.2.2
neo4j==1.7.6
dataclasses-json==0.5.2
marshmallow==3.4.0
structlog==19.1.0
newrelic>=5.0.0
more-itertools==8.0.2
mimesis==3.3.0
pyhumps==1.3.1
python-dateutil
typing-extensions==3.7.4
boto3==1.11.1
